Redesign with Abstract SCMSource
