+++ src/main/resources/org/jenkinsci/plugins/p4/scm/BranchesScmSource/config-detail.jelly
@@ -1,0 +1,27 @@
+<?jelly escape-by-default='true'?>
+<j:jelly 	xmlns:j="jelly:core" 	xmlns:st="jelly:stapler" 	xmlns:d="jelly:define"
+			xmlns:l="/lib/layout" 	xmlns:t="/lib/hudson" 		xmlns:f="/lib/form"
+			xmlns:c="/lib/credentials">
+
+    <f:entry title="Perforce Credentials" field="credential">
+        <c:select/>
+    </f:entry>
+
+    <f:entry title="${%Include streams}" field="includes">
+      <f:textbox default="//depot/..."/>
+    </f:entry>
+
+    <f:entry title="${%Character Set}" field="charset">
+    	<f:select default="none"/>
+    </f:entry>
+
+    <f:entry title="Workspace Name Format" field="format">
+        <f:textbox default="${descriptor.defaultFormat}"/>
+    </f:entry>
+
+    <f:entry title="Populate options">
+        <f:dropdownDescriptorSelector field="populate"/>
+    </f:entry>
+
+    <t:listScmBrowsers name="p4.browser"/>
+</j:jelly>
